+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v0.7.1...main
+[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v0.7.2...main
+
+## [Unreleased][]
+
+### Added
+
+- Added `sqlitex.Transaction`, `sqlitex.ImmediateTransaction`, and
+  `sqlitex.ExclusiveTransaction`.
 
 ## [0.7.2][] - 2021-09-11